+++ src/decoding.rs
@@ -12,7 +12,7 @@ use symphonia::core::formats::probe::Hint;
 use symphonia::core::formats::{FormatOptions, FormatReader, TrackType};
 use symphonia::core::meta::MetadataOptions;
 
-pub(crate) fn decode_media_data<R: std::io::Read + Send + Sync + 'static>(
+pub(crate) fn decode_media_data<R: std::io::Read + Send + Sync>(
     input: R,
     target_sample_rate: f32,
 ) -> Result<AudioBuffer, Box<dyn std::error::Error + Send + Sync>> {
@@ -94,20 +94,20 @@ impl<R: Read + Send + Sync> symphonia::core::io::MediaSource for MediaInput<R> {
 /// Media stream decoder (OGG, WAV, FLAC, ..)
 ///
 /// The current implementation supports Symphonia's audio formats and codecs.
-pub(crate) struct MediaDecoder {
-    format: Box<dyn FormatReader>,
+pub(crate) struct MediaDecoder<'a> {
+    format: Box<dyn FormatReader + 'a>,
     decoder: Box<dyn AudioDecoder>,
     track_index: usize,
     packet_count: usize,
 }
 
-impl MediaDecoder {
+impl<'a> MediaDecoder<'a> {
     /// Try to construct a new instance from a `Read` implementer
     ///
     /// # Errors
     ///
     /// This method returns an Error in various cases (IO, mime sniffing, decoding).
-    pub fn try_new<R: std::io::Read + Send + Sync + 'static>(
+    pub fn try_new<R: std::io::Read + Send + Sync + 'a>(
         input: R,
     ) -> Result<Self, Box<dyn std::error::Error + Send + Sync>> {
         // Symphonia lib needs a Box<dyn MediaSource> - use our own MediaInput
@@ -188,7 +188,7 @@ impl std::fmt::Display for UnsupportedAudioCodecError {
 
 impl std::error::Error for UnsupportedAudioCodecError {}
 
-impl Iterator for MediaDecoder {
+impl Iterator for MediaDecoder<'_> {
     type Item = Result<AudioBuffer, Box<dyn Error + Send + Sync>>;
 
     fn next(&mut self) -> Option<Self::Item> {